Safari-609.1.11.1/LayoutTests/http/tests/contentextensions/popups.html

<html> 
<body>
<script> 
if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.setCanOpenWindows(true);
    testRunner.setCloseRemainingWindowsWhenComplete(true);
}

win = window.open("about:blank");
alert(win ? "FAIL" : "about:blank successfully blocked");
win = window.open("data:text/html, hello");
alert(win ? "data url should not be blocked" : "FAIL");
win = window.open("data:text/html, about");
alert(win ? "data url should not be blocked" : "FAIL");
</script> 
</body> 
</html>
