Safari-609.1.11.1/LayoutTests/http/tests/contentextensions/popups-expected.txt

CONSOLE MESSAGE: line 10: Content blocker prevented frame displaying http://127.0.0.1:8000/contentextensions/popups.html from loading a resource from about:blank
ALERT: about:blank successfully blocked
ALERT: data url should not be blocked
ALERT: data url should not be blocked
